Obey browser language preferences
Before this change, the site would always be displayed in Hungarian
until the user explicitly set their language preferences in the menu.
This is inconvenient for guests who do not speak the language; I
personally had to had to help a guest sign up, as they did not notice
the hamburger menu on their phone.

This change makes the `Locale` middleware select the appropriate
language via the `Accept-Language` header. It can still be overridden
via the Language menu.