Use the main navbar layout on the welcome page
This automatically turns the top nav links ("Login", "Application for
college students" and "Register for tenants") automatically turn into a
sidenav on mobile.

The easiest way I could make the logo stay centered on desktop is to
make the side nav only show on mobile layouts. I made the navbar blade
take a parameter indicating if we want it fixed (like on the main UI).

The layout of other pages are unaffected.