Always prefer Hungarian if the user speaks it (#585)
* Always prefer Hungarian if the user speaks it

Even if the user has English as the highest priority entry in their
Accept-Language header, we should select Hungarian by default if it's
one of the languages that the user speaks.

* Always show the Collegist registration form in Hungarian

Even if the user's language is English, Collegist registration will now
display in Hungarian. To prevent confusion among foreign guests, the
"This form is for the prospective collegists. Click here to register as
a guest resident." message is displayed in English alongside Hungarian.