Remove large, commented out script
Even though the code never gets executed as it's wrapped in JavaScript
comments, the blade templating engine still processes it and emits it in
the output. In some pages related to the general assembly, this causes
the names of all users and their user IDs to be included in the page
source as comments.

We could use blade comments (`{{-- [...] --}}`) if we wanted to keep
this code around but not include it in the output; but it serves no
purpose and can be restored from git history if it ends up being needed.

Fixes #514