Fix lp command for one-sided printing
We were including an empty string in the arguments of `lp` when
performing one-sided printing. This was interpreted as an additional
file to be printed at path "". Also, this API leaves argument splitting
to us; how did it even work with `-o` and `sides=two-sided-long-edge`
being parsed as a single argument?