Remove string processing done via shell and #486
Using shell scripts to transform text contributes quite a bit to the
smell of the codebase. We also had issues because of this (`grep`
arbitrarily decided sometimes for outputs of `pdfinfo` that they were
binary data and not text). While removing calls to
`Process::fromShellCommandline` (which is quite unsafe), I also moved
from `ping` to `nmap` (fixing #486).