Fix GHSA-59cf-4f4p-752w
KKT handlers (and similar) had permission to manage anonymous questions and community service.
The code sometime only checked for the STUDENT_COUNCIL role and did not check if the user was an elected member of the student council. STUDENT_COUNCIL role is given to many people (including KKT handlers, committee members).

This commit should fix these issues, a warning is also placed in
`Role.php` to decrease the likelihood of making this mistake again.