Fixing #653
I checked the issue. The main problem is that using livewire with dates seems to be quite impractical.
Like: putting a simple text field `<input type="text" value="{{$firstDayYYYYMMDD}}">` works fine, but updating does not work (at all) if we navigate back to the page with dates on Chrome: `<input type="date" value="{{$firstDayYYYYMMDD}}">`
With `<input type="date" wire:model="firstDayYYYYMMDD">` and `<input type="text" wire:model="firstDayYYYYMMDD">` updating works even if we navigate back to the page but it shows the old date for the first time in Chrome.
I'm not an expert in Livewire and from all of my tried solutions, I could only get it to work by waiting for a bit and update the value of the input field.

I could not get it to work with materialize's datepicker on chrome, so the original HTML element is used.

I am quite unhappy with the quality of this change but at least it makes it functional.