Added texts that can be configured from the UI
Texts on some places during the application process were edited multiple times last year which was not ideal as they were hard-coded. And we also had an issue if we needed to display information based on which workshop the user applied to.
In this commit I extended the functionality of the `custom` table, renamed it to `configurable_texts` which can be used for storing text for keys and potentially workshops. I added also an admin interface where these settings can be changed.