Fix Alfonso tests for the spring semester
The function that checks whether a user has been admitted as a bachelor
student assumes that admittance semesters are autumn semesters. This is
a fine thing to do in my opinion, but the tests didn't account for that
and therefore were broken when we ran them in the spring.

This resolves the two unit test failures on CI.

Fixes #629